move aws and google kms together in mod
